Extract Dsl and Hook from Serializable module
