+++ lib/pb/serializer.rb
@@ -3,6 +3,8 @@ require "the_pb"
 require "computed_model"
 require "google/protobuf/field_mask_pb"
 
+require "pb/serializer/dsl"
+require "pb/serializer/computed_model_hook"
 require "pb/serializable"
 require "pb/serializer/base"
 require "pb/serializer/attribute"
